Build and link Box2D in the build script.
Build and link the Box2D submodule in build.rs, not requiring users to have Box2D installed on their system.
This significantly reduces the effort to use this library on Windows, since it does not have a package manager to install Box2D.
This does require users to have CMake installed, because this is what Box2D uses as its build system.